Separate refcount into "by threads" and "by world objects" parts.
This is needed because from the POV of the GC running on a thread,
a reference due to an object cycle and a reference by another
thread looks exactly the same.

The RC delta by threads is not a power of two. It would look very
professional, but we only add and subtract from the refcount so
that is more than enough.